Add stubs for 'About' page and alternate layouts
In order to add the 'About' without a Jumbotron while keeping the
Jumbotron for the 'Home' page, it's easiest to set the default layout to
add the page content to a container (for some nice margins), and create
an alternate layout without the margins to keep content fluid.

In order to use the should-matchers test for layouts we added
rails-controller-testing to the project.

The route specs for the root route are separate from the rest of
PageController routing specs. This will be useful as more complicated
root routing is implemented (such as for logged in vs guest users).