Add Username to User
This is the first step in adding user profiles to users. Because
usernames potentially be used in lieu of email addresses when logging
in, we add them to the user model directly rather than to a separated
profile object (which will contain additional profile data that is not
used in the session management process).