Add Item model and create/index/show actions
Added the Item model and controller actions for creation, show, and
index. As of this commit, Item creation is not tied to a specific user
or, indeed, limited to logged in users. User authorization and
update/delete actions will come in the next commit, and further item
business logic in one after that.