Add association between Users and Items
Users now own the items they create, and only the owner can see their
items in an index or show action. When an anonymous user attempts to
view an Items action they are redirected to the login page. When a
logged in user attempts to use the #show action on an Item they do not
own, they are shown an alert message and redirect to their Item index.