Move text to locales and extract form partial
A final refactor before closing this feature is to move all static text
to a locales factor, to follow best practices for internationalization.

Also, since the new and edit forms are pretty much identical, they are
extracted to a shared `_form` partial, with the view spec moved to that
as well.