Add extra fields to Items
Flesh out the Item model by adding extra fields, as returned by the ISBN
lookup. These fields are:

* authors
* categories
* description
* publisher
* published_on
* page_count

Some of these fields (:authors, :categories) are stored as
Postgres arrays. The #authors= and #categories= methods are modified to
accept either strings or arrays and split strings appropriately when
used (e.g. accepting a comma-separated list from a web form). The form
itself converts these arrays back to comma-separated strings for easier
editing. Note that this sacrifices the ability to use commas in the
values themselves (unless posting via JSON API). This might be addressed
in the future using JavaScript to dynamically update the form.