Refactor user interface
Refactored the user interface for a bit more consistency between
templates for Items. Added the "Lookup by ISBN" button to an input group
for a cleaner look, and made a hidden default "Submit" button to the
user can hit `Enter` to submit a normal, manually entered item.

Also, a bug when lookup on a blank ISBN (or really, any ISBN lookup that
did not return both an ISBN-10 and ISBN-13) was fixed.

A 'bulk entry' form with just an ISBN lookup should come later.