Users can filter by category
By clicking on a category tag on the show page, users can show an index
of items matching that category.

The scope looks for one exact match within the Item's categories array.
An additional scope is added for partial, case-insensitive matches,
along with parameters on the index route, but no UI for this is
currently implemented.